Whole shell in a single window with a transparent background.
Creating a window for Panel and another one for Launcher is
complicated. Let's see if with a single window we can get the
desktop shell ready without having to worry about the focus on the
compositor side. Focus has problems because on Wayland, just like
it was on X11, only one window at a time has the focus and grabs
the keyboard.