Update due to changes in the wayland-cursor API
Greenisland doesn't seem to use mouse cursors yet, but now that cursors
are wl_surfaces, that means there is no longer a 1:1 connection between
surfaces and windows.  Before creating WaylandSurfaceItems from surfaces
we need to make sure that the surface is a window surface (by checking
for a wl_surface_shell).