Disable compositors build make the shell working with Weston
QtCompositor and kms support are immature and Qt 5.1 is focused on
getting the QPA plugin done, so compositor improvements are scheduled
for Qt 5.2, in fact the QtCompositor library is not built by default
because it's an experimental API.

This leaves me no choice but port the desktop shell to Weston.

Compositors are excluded from the build.

Issue: #23