React to background changes with animations
We create a rectangle for solid background, one for gradients and an
image for wallpapers and when the background is changed we fade between
the previous background and the new one.

In case an invalid background is chosen (i.e. editing by hand the
configuration file writing an invalid value) the background goes black.