Remove popup surfaces from model when unmapped
Popup menu surfaces (at least with Qt applications) are not destroyed so
their element is not removed from the surface model and the next time
the same menu is open, surfaceMapped() sees it and thinks it doesn't have
to create a window representation.

But we get an unmapped signal so it's there that we remove the surface
from the model and destroy its representation.