Use KScreen instead of our own solution
KScreen is better than QScreen as it's capable of detecting when primary
screen changes and has a nice fake backend.
We'll need to add Wayland support to KScreen though.

The initialization sequence was changed a little bit to get the screen
estate in advance and change compositor output geometry.
Also, the screen model is now created from main().