Say hello to Green Island launcher
The launcher is a frontend to Green Island that prepares the environment
for it by setting the appropriate variables and passing the right
arguments.

This is useful on devices such as Raspberry Pi that needs particular
QPA plugins or on KMS where we need to run a Weston compositor with
the wl_fullscreen_shell interface because the kms plugin won't work.

Having all this logic in Green Island itself would be a problem
because we need to set environment variables before QGuiApplication
instance is created but we need to parse the command line argument
with QCommandLineParser which requires the application instance.
Also, this separates things which is always good because each
process deals with its own set of problems.