Add fullscreen shell support (client side)
When Green Island acts as a Wayland client it needs to be nested into
another compositor that advertise the wl_fullscreen_shell interface
such as Weston.

Currently the only use for that is when running on KMS, Weston has a far
better support for it compared to the QPA plugin. The QPA plugin also
need logind integration.  Hopefully one day Green Island will not need
another compositor for KMS.

This come in handy for other platforms where KScreen doesn't have a
backend available.  A Wayland backend will always be available so it's
always a nice fallback.

On other platforms Green Island won't need to be nested except when a
KScreen backend is missing.