fullscreen-shell: Don't force a particular mode
When Green Island is nested into another compositor we'll use the QScreen
backend for KScreen which means that the outputs configuration will be
picked up from the other compositor.

Being the output configured the same we don't need to ask for a different
mode, the other compositor is already good to go.