Add splash surface role and rearrange layers
Some desktops like Plasma have a little program that displays a
splash screen so we add a specific role for its surfaces.

We rearrange the layers and now the black rectangle used for fade-in
and fade-out effects is now a layer for splash screens.

The fade effect still remains so that desktops without a splash screen
are still supported.  As a nice side effect this will also make
Plasma splash screen go away with style.