Load external plugins
Install a Plasma package definition and add a command line argument to
load a different compositor plugin.

If no plugin is specified we keep loading internal code from resources,
however at some point we will move a part of this code to Hawaii desktop
and to the Green Island plugin for QML.

Later, we might want to move plugin loading code where appropriate.

This change adds a dependency to the Plasma framework, this is undesired
and so it will be changed as soon as the KPackage framework is ready.