plasma: Don't keep destroyed surfaces in the list
Problem: the compositor bails out with an error of surface role already
assigned if we run plasmashell, close it and run it again because the
Plasma surface is always in the list even when destroyed.

Destroy Plasma surface when the actual surface is destroyed

This will also trigger Plasma surface destroy signal that we catch to
remove the surface from the list.

Removing the surface from the list fix the problem.