Always use QScreen backend when fake data is not provided
This is the only sane backend we can use.

In the future, when there will be a Wayland backend, we'll be able to
use that when the compositor runs nested into another compositor.
Even in this case, QScreen would still be a good choice though because
it would get the screen configuration from the Wayland QPA plugin.