Use xcb_egl integration when running on X11
The xcb platform plugin has seen a number of improvements with Qt 5.5,
one of which is separate GL integrations for xcb.

Without this the compositor won't be able to use EGL which means that
when using Qt 5.4 it won't get access to the EGL display.

The most visible effect of this is Qt clients without decorations.