wl-shell: Do not move full screen windows
This will fix double click on mpv.

mpv moves itself when clicked, this a double click will trigger a move
operation and then set the surface full screen before the move grabber
is destroyed.

Since our move grabber set the position we have this was causing the
window position to be recalculated.

Exit the move grabber if the window is full screen to avoid
the problem.