Be smarter determining the app_id
Add more heuristics to determine the app_id.

First of all check whether the application provides such information with
the gtk-shell protocol (only for GTK+ applications), then use the app_id
from xdg-shell or wl-shell and try to prepend a vendor prefix in an
attempt to support Hawaii and the major desktops. If all the above fails
use the executable base name.