Remove libkscreen dependency, part 1
libkscreen recently introduced a problematic change: now it loads
backends with an external process that communicates via D-Bus.

This caused a lot of problems to Green Island: crashes on exit,
slow startup, random failure on getting the screen configuration.

Since Green Island only used the QScreen backend we can easily
replace that with a simple backend for our ScreenManager class
that we call "native", resulting in the complete removal of code
that maps KScreen::Output to our Output class which is now
directly used by the screen manager.

Green Island is now back to being software that only depends
on Qt.

Next commit will introduce a backend for user configurations.