Create a client connection to the compositor itself
This is needed for cursor theme support, we want to use the wayland-cursor
API to retrieve cursors for move and resize operations however since
cursors are stored in a surface it relies on the shared memory pool
which can only be created by clients.

In order to do that we create an internal client connection to
ourselves and bind wl_shm creating a shared memory pool.

Issue: #73