Remove decorations from output windows only when using the native backend
With the fake screen backend the user will craft a JSON configuration that
will fit his setup. This configuration will likely be smaller than the
screen geometry hence the title bar comes in handy to move the window.

On the other hand with the native backend on X11 we don't really want
title bars as they take space out and makes reaching the bottom part
of a shell nearly impossible.

See #93