compositor: Socket name argument
Rename --socket to --wayland-socket-name which is handled internally
by QtCompositor.

Since Compositor is a singleton we create it without a socket name,
then QWaylandCompositor will see --wayland-socket-name and set it
on its own.

This is ugly and need to be changed on QtCompositor but meanwhile
we need to do this.