Import eglfs, eglconvenience and fbconvienience from Qt
A mix between 5.5 and dev branches as of today.

Since the eglfs refactor and the introduction of the kms hooks
with Qt 5.5 a few fixes have been pushed to the 5.5 branch plus
a number of features will be developed for Qt 5.6 in the dev
branch.

We need a custom platform plugin because we cannot wait for
a new Qt version to pick up fixes and new features (and we know
that eglfs still needs some work) besides not all distributions
will likely update their Qt packages when we need and our release
schedules do not necessarily match.

There's also another reason: eglfs support for ARM boards will
require qtbase to be built with the vendor GL libraries which
means rebuilding qtbase for each board and that's tough.

However Green Island is way faster to build and everyone can
deal with multiple builds, leaving the distro qtbase untouched.

I do not like forks so this code should be kept in sync with
Qt as much as possible.

Issue: #100