Import QtCompositor API here
The QtCompositor API is not stable and we have a hard time
following it.

Our release schedule doesn't often match the QtCompositor one,
hence we find ourselves using its dev branch and even apply
patches on top of it that cannot be merged upstream.
One notable example is Wayland 1.6+ support which cannot be
applied because CI supports only Wayland 1.4, but this patch
is vital for us otherwise Gtk+ applications won't work.

This means that Hawaii constantly need a patched qtwayland version
and that makes packaging harder.

We need a QtCompositor copy here until the new QtWaylandCompositor
stable API is ready and shipped with a Qt version we can depend on.

QtWaylandCompositor will be shipped as a technical preview with
Qt 5.6 and it will likely be finalized for Qt 5.7 or 5.8, however
we'd like to be compatible with Qt 5.5 for a while.

Once the API is complete, stable and shipped by most distributions
we will drop our fork.

Issue: #104