Build only EGL device integration plugins
Our goal here is to help binary distribution of Green Island.

A distro would build the Green Island sources the first time in order
to have the libraries and tools packaged, then build a second time to
build the a specific integration so that each device integration is
in its own package.

The packages layout would be like:

- greenisland
- greenisland-devel
- greenisland-kms
- greenisland-brcm
- greenisland-mali
- greenisland-viv
- greenisland-x11