server: Refactor HomeApplication
Move the code to a subdirectory and refactor.

This class locates metadata.desktop for the specified compositor
identifier (from the data directory), reads the file and tries to load
the master QML file.

It internally uses QQmlApplicationEngine to load the main file.

Setting the screen configuration file will make the ScreenManager
component use the fake screen backend.
This property is irrelevant for compositors not using the
ScreenManager component.

HomeApplication doesn't setup the tty anymore, that is now done by
the greenisland QPA plugin.

Before running the QML master file, it verifies whether the directory
referenced by the XDG_RUNTIME_DIR environment variable is available
and has the right permission.  Execution fails with an error if
the directory doesn't meet the requirements.

With this commit we also move more files to the core subdirectory.

Issue: #131