Test QML shell for xdg_shell
Add a QML shell to test the xdg_shell extension without
all the bells and whistles of a real shell.

We want to make an abstraction for different desktop
grade shells (wl_shell, xdg_shell and XWayland) in order to have
the same window management behavior over different shells.

With a xdg_shell only implementation we can test just this
extension while with the simple shell we will see the unified
shell in action.

Issue: #129