declarative: Desktop window management
This component initializes wl-shell, xdg-shell and gtk-shell and
manages windows.

With each extension we create specialized shell surfaces that
are reparented to a unified window component that implements
animations for mapping and destruction of toplevel, transient
and popup windows.

The window component exposes typical shell surface properties such
as title and appId, including the state offering a unified API.

Issue: #77