server: Introduce QuickOutput
This class is an evolution of the old Output class, the name reflects
the fact that it's meant for the QtQuick world.

The QuickOutput class uses the new QtWaylandCompositor API.

It keeps properties for the hotspots and sets the window screen
during the initialization, this is necessary for the eglfs and
greenisland QPA plugins because they mandate a window for each
screen but windows are all assigned to the primary screen by
default.

When a window is assigned from QML it must NOT be set to visible,
otherwise the platform window will be created before initialization
preventing us from setting the screen.
QuickOutput will automatically set the window visible upon
initialization.

QuickOutput is also ready for the screencaster protocol.
When the extension will be updated to the new API we'll just have
to fill the slot connected to the afterRendering signal with
a call to take a picture of the current frame.

Issue: #128