server: Window manager
Rewrite the window manager component with C++.

Window manager is no longer a compositor, it's a component that needs a
compositor that already exist.

It creates wl-shell, xdg-shell and gtk-shell extensions and creates
the shell surfaces and one QtQuick view for each output.

ClientWindow acts as a proxy for shell surface specific properties.

ClientWindowView is the main QtQuick view and reparents the protocol
specific surface item.  It handles input events triggering the ping
request, move dragging in any place holding the modifier key and in
the future it might draw decorations.

Closes #146