platform: DISPLAY must not be empty in order to load the x11 plugin
The DISPLAY environment variable could be set to an empty string so
we better check whether it is empty or not.

This might be the case when running the compositor on a terminal and
for some reason we want to skip the x11 plugin, to do so we pass
the DISPLAY empty.