platform: Activate primary screen window
This makes keyboard input work.

Wayland windows are created on the primary output so input will be
processed there and sent to the first view, all shell UI is usually
created only on the primary output (the decision is up to the shell but
usually that's the way it works) so with the window being active, text
fields will have focus there.