client: Fixes to Buffer
Remove the public constructor because nobody is going to create a buffer
without going through ShmPool. This also fixes going from wl_buffer to
Buffer inside CursorTheme, as creating a new Buffer is just wrong we
need to get the existing instance from the user data instead.

Fixes to the format decoding and adding a new RGBA32 format.