Add mode support to QWaylandOutput
Outputs usually have more than one mode, add an API to support them.

When sizeFollowsWindow is true a fake mode the size of the window is
added and with private slots we track window size changes and modify
the mode to accomodate the window size.  However sizeFollowsWindow
is no longer true by default, a compositor will have to explicitly
choose between adding modes or setting the property to true.

The setGeometry() method is gone as it doesn't make sense now, the
setWidth() and setHeight() methods are now private slots that resize
the fake mode when sizeFollowsWindow is true.

Adapt the unit test to these changes.