waylandcompositor: Proper fix for buffer destruction
The previous fix (93ca929fb9caf347150) would always crash on client exit.
It tried use the same destroy_listener on several buffers. Due to the
way linked lists are implemented in Wayland, a listener can only
be listening to one signal at the time.
The correct way to do this is to create a new listener for each buffer.

Done-with: Paul Olav Tvete <paul.tvete@theqtcompany.com>