waylandcompositor: Rename compositor Shell to WlShell
 * QWaylandShell -> QWaylandWlShell
 * QWaylandShellSurface -> QWaylandWlShellSurface
 * QWaylandQuickShellSurfaceItem -> QWaylandQuickWlShellSurfaceItem
 * Shell -> WlShell (QML)
 * ShellSurface -> WlShellSurface (QML)
 * ShellSurfaceItem -> WlShellSurfaceItem (QML)

This is done to avoid confusion with XdgShell and will hopefully help
clarify that some of the examples only support wl_shell and not
xdg_shell.

Additionally, this makes "Shell" an available name in the compositor
API, which may in turn enable the creation of a more general abstraction
hiding the details of the shell backends (i.e. xdg_shell, wl_shell, and
eventually ivi_shell).

Done-with: Johan Helsing <johan.helsing@theqtcompany.com>