Remove unused signal QWaylandSurface::pong
It appears to be a leftover from when QWaylandSurface::ping existed
(which forwarded the ping event to the wl_shell interface for the client
corresponding to the wl_surface).

There is no guarantee that the wl_surface has a role that supports some
form of ping, so it doesn't make much sense to keep the signal.

Done-with: Johan Klokkhammer Helsing <johan.helsing@theqtcompany.com>