Add initial XWayland QML plugin
Start porting the old code to the new API as a QML plugin.
This will never be complete in time for 0.8 but I want to do
some work ahead.

It barely work with GNOME applications and Spotify, you can see
and move windows that is.

The QtQuick integration is modeled after the new QtWaylandCompositor
API in that there is a shell surface , a QtQuick representation
and an integration logic class.

The X11 window manager creates the shell surface and assign a
surface to it the moment it is available, then a signal is emitted
so that the compositor creates the QtQuick representation that in
turn creates the integration module.

When an event such as move and resize is received by the X11
window manager a signal from the shell surface is emitted and
received by the integration module that performs the action.