deviceintegration: Delay initialization until logind is ready
For a device integration that needs logind we wait until we are
connected and the session control has been acquired before doing
the initialization.

Once the initialization is finished we emit a signal that is
connected to the event loop created in initialize() so that
the method delete the event loop and returns.

If the device integration doesn't need logind we preceed with
the synchronous initialization as before.

Additionally, the native interface is now separate and the
platform integration is a QObject so it can emit the initialized()
signal when the actual initialization is done.